refactor: Make anonymous classes into actual test asset classes
Simplifies test definitions.
